+++ src/index.ts
@@ -3,7 +3,7 @@ import app from './lib/app';
 export { middleware } from './lib/middleware';
 export { combineReducers } from './lib/reducers';
 export { actionFlat, actions, actionValidate } from './lib/utils/store';
-export { bindActions } from './lib/utils/connect';
+export { memo } from './lib/utils/connect';
 
 export { DispatchType, UnpackedArray, ActionReturnType, ActionType } from './lib/types';