+++ src/lib/utils/connect.ts
@@ -1,5 +1,4 @@
-import { ActionSubjectType, StoreDispatchType } from '../types';
-import { hasProperty, isObject } from './helper';
+import React, { FunctionComponent } from 'react';
 
 export const defaultMapStateToProps = (): any => {
   return {};
@@ -17,52 +16,13 @@ export const mergeProps = <MapState, MapDispatch, ConnectedProps, ComponentProps
   return ({ ...props, ...stateProps, ...dispatchProps } as any) as ComponentProps;
 };
 
-export const bindActions = <State, T extends Record<any, (...args: any[]) => ActionSubjectType<State, any>>>(
-  objs: T,
-  dispatch: StoreDispatchType<State, any>,
-): T => {
-  const data: any = {};
+export const memo = <C, W>(Component: FunctionComponent<C>, Wrapper: FunctionComponent<W>) => {
+  const displayName = Component.displayName || Component.name;
 
-  Object.keys(objs).forEach((key) => {
-    data[key] = (...args: any[]) => dispatch(objs[key](...args));
-  });
-
-  return data as T;
-};
-
-export const isStrictEqual = (a: any, b: any): boolean => {
-  return a === b;
-};
-
-export const isEqual = (a: any, b: any): boolean => {
-  if (isStrictEqual(a, b)) {
-    return a !== 0 || b !== 0 || 1 / a === 1 / b;
-  } else {
-    return a !== a && b !== b;
-  }
-};
-
-export const shallowEqual = (a: any, b: any): boolean => {
-  if (isEqual(a, b)) {
-    return true;
+  // istanbul ignore else
+  if (displayName) {
+    Wrapper.displayName = displayName;
   }
 
-  if (!isObject(a) || !isObject(b)) {
-    return false;
-  }
-
-  const keysA = Object.keys(a);
-  const keysB = Object.keys(b);
-
-  if (keysA.length !== keysB.length) {
-    return false;
-  }
-
-  for (const key of keysA) {
-    if (!hasProperty(b, key) || !isEqual(a[key], b[key])) {
-      return false;
-    }
-  }
-
-  return true;
+  return React.memo(Wrapper);
 };