+++ src/lib/utils/equals.test.ts
@@ -1,73 +1,7 @@
 /* tslint:disable:no-null-keyword */
-import {
-  bindActions,
-  defaultMapDispatchToProps,
-  defaultMapStateToProps,
-  isEqual,
-  isStrictEqual,
-  mergeProps,
-  shallowEqual,
-} from './connect';
+import { isArrayEqual, isEqual, isStrictEqual, shallowEqual } from './equals';
 
-describe('Check the connect functions', () => {
-  test('it should be return an object when defaultMapStateToProps() is called', () => {
-    const returnValue = defaultMapStateToProps();
-
-    expect(typeof returnValue).toBe('object');
-    expect(returnValue).not.toBeNull();
-  });
-
-  test('it should be return an object when defaultMapDispatchToProps() is called', () => {
-    const returnValue = defaultMapDispatchToProps();
-
-    expect(typeof returnValue).toBe('object');
-    expect(returnValue).not.toBeNull();
-  });
-
-  test('it should be return the merged object when mergeProps() is called with empty objects', () => {
-    const returnValue = mergeProps({}, {}, {});
-
-    expect(typeof returnValue).toBe('object');
-    expect(returnValue).not.toBeNull();
-  });
-
-  const testFunction = (a: any) => a;
-  const testTrue: Array<[object, object, object]> = [
-    [{ text: 'string1' }, { mode: 1 }, { text: 'string1', update: testFunction, mode: 1 }],
-    [{ text: 'string1' }, { text: 'string2' }, { text: 'string1', update: testFunction }],
-  ];
-
-  test.each(testTrue)(
-    'it should be return the merged object when mergeProps() is called [%p, %p]',
-    (stateProps, ownProps, expected) => {
-      const returnValue = mergeProps(stateProps, { update: testFunction }, ownProps);
-
-      expect(returnValue).toEqual(expected);
-    },
-  );
-
-  test('it should be return object with dispatch funcs when bindActions() is called', () => {
-    const dispatch = jest.fn();
-
-    const actions: any = {
-      add: (id: number, message: string) => ({ type: 'add', id, message }),
-      del: (id: number) => ({ type: 'del', id }),
-    };
-
-    const result = bindActions(actions, dispatch);
-
-    expect(Object.keys(result)).toEqual(['add', 'del']);
-    expect(result.add).toEqual(expect.any(Function));
-    expect(result.del).toEqual(expect.any(Function));
-
-    result.add(1, 'message');
-    result.del(1);
-
-    expect(dispatch).toHaveBeenCalledTimes(2);
-    expect(dispatch).toHaveBeenNthCalledWith(1, { id: 1, message: 'message', type: 'add' });
-    expect(dispatch).toHaveBeenNthCalledWith(2, { id: 1, type: 'del' });
-  });
-
+describe('Check the equals functions', () => {
   describe('Check the function isStrictEqual()', () => {
     const testObject = { a: 5, m: 'test' };
     const testStrictTrue: Array<[any, any]> = [
@@ -126,11 +60,53 @@ describe('Check the connect functions', () => {
     });
   });
 
+  describe('Check the function isArrayEqual()', () => {
+    const testObject = { a: 5, m: 'test' };
+    const testEqualTrue: Array<[any, any]> = [
+      [[], []],
+      [
+        [1, 3],
+        [1, 3],
+      ],
+    ];
+    const testEqualFalse: Array<[any, any]> = [
+      [1, 10],
+      [-1, 0],
+      [[1], []],
+      [
+        [1, 5],
+        [5, 1],
+      ],
+      [{ a: 3 }, { a: 3 }],
+      ['test', 'test'],
+      [testObject, testObject],
+      [null, null],
+      [undefined, undefined],
+      [NaN, NaN],
+      [true, false],
+      [true, true],
+      [false, false],
+    ];
+
+    test.each(testEqualTrue)('isArrayEqual(%s, %s) is true', (a, b) => {
+      expect(isArrayEqual(a, b)).toBeTruthy();
+    });
+
+    test.each(testEqualFalse)('isArrayEqual(%s, %s) is NOT true', (a, b) => {
+      expect(isArrayEqual(a, b)).toBeFalsy();
+    });
+  });
+
   describe('Check the function shallowEqual()', () => {
     const testShallowObject = { a: 5, m: 'test' };
     const testShallowTrue: Array<[any, any]> = [
       [1, 1],
       [0, 0],
+      [[], []],
+      [
+        [1, 3],
+        [1, 3],
+      ],
       ['test', 'test'],
       [testShallowObject, testShallowObject],
       [{ a: 3 }, { a: 3 }],
@@ -144,6 +120,11 @@ describe('Check the connect functions', () => {
     const testShallowFalse: Array<[any, any]> = [
       [1, 10],
       [-1, 0],
+      [[1], []],
+      [
+        [1, 5],
+        [5, 1],
+      ],
       [{ c: { a: 1 } }, { c: { a: 1 } }],
       [{ c: { a: 1 } }, { c: { a: 1 }, d: true }],
       [true, false],